[NFC] Distinguish between getModelOptPath and getModelOptDir
For some reason, model path in current implementation always expects to be a
path to a directory, while it is not true. It can be a file as well. This
patch changes logic of getModelOptPath so that it can return either path to
model directory or model file. For old functionality, getModelOptDir is used.