Introduce parallel minibatch processing in Image Classifier
This patch allows Image Classifier to process minibatches in parallel.
When running inference with -minibatch option, you can also specify the
number of threads by passing the -minibatch-threads option. It will create
a thread pool of the specified size, and each minibatch will be inferred
independently by a thread from this pool.